sensors/requirements.txt
# Python dependencies for sensors
paho-mqtt==1.6.1
aiocoap==0.4.7
cryptography==41.0.7
asyncio-mqtt==0.13.0